Refactoring

+++ src/main/java/model/Cannon.java
@@ -3,8 +3,8 @@ package model;
 import java.awt.*;
 
 public class Cannon extends Rectangle.Double {
-    public static final int WIDTH = 50;
-    public static final int HEIGHT = 50;
+    public static final double WIDTH = 50;
+    public static final double HEIGHT = 50;
     private boolean team;
 
     public Cannon(double x, double y, boolean team){
